Avoid use of global variables
While trying to make single section binary using llvm-mingw, ld.lld revealed a bug that can
be avoided by not use of global variables so let's just play along and avoid use of global
variables as the formation of app_state_t has resulted in code clarity also.

And this single section binary made by llvm-mingw now has the same size as the unchanged
MSVC generated binary size, 5632 bytes, and now we have two set of compiler toolchains
generating the same binary size.

Apparent binary size of MSVC artifact isn't changed here but looking more carefully
at the binaries now there is less padding in .code and .idata sections of the MSVC
generated binary so there were some binary size increase (should be mainly due to new
Win32 APIs that are called) but just not an apparent change.

This however raises the support of the app to Windows XP from the already broken
Window 2000 support as SecureZeroMemory doesn't exist in Windows 2000 (I suppose?)
but was needed to avoid MSVC's tendency to insert a memset when a local variable
needs to be zero initiated. Guess it can be avoided but I think it doesn't worth
the hassle given the tray couldn't be seen in Windows 2000 already and the MSVC
generated didn't support Windows 2000 (but the mingw one did which wasn't on the
release anyway).